+++ CHANGELOG.md
@@ -28,6 +28,7 @@ All notable changes to this project will be documented in this file.
 - *(cards)* Saved locally on browser encrypted
 - *(cards)* Fixed the load from cache and class refactor redundance
 - *(cards)* Created the card topics pagination
+- *(cards)* Created the card topics pagination for both mobile and desktop
 
 ### ⚙️ Miscellaneous Tasks